Address the issue relating circular dependency

+++ main.cpp
@@ -1,4 +1,5 @@
 #include "vector.hpp"
+#include "particle.hpp"
 
 int main()
 {
@@ -9,9 +10,9 @@ int main()
 
     for(int i=15; i>=0; i--)
     {
-        particle->dragForce->updateForce(particle,0.001);
+        particle->dragForce->updateForce(particle, (float)0.001);
         particle->posUpdate(particle->velocity,
-                            particle->acceleration,0.001);
+                            particle->acceleration,(float)0.001);
 
         std::cout<<particle->velocity.x <<" "<<particle->velocity.y
                  <<particle->velocity.z<<'\n';
